push final project

+++ routes/api.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\ProductController;
 use App\Http\Controllers\UserProductController;
 use App\Http\Controllers\OrderController;
 use App\Http\Controllers\OrderDetailsController;
+use App\Http\Controllers\AdminController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -28,14 +29,30 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::resource('/users',UserController::class);
 Route::post('/users/login',[UserController::class,'login']);
-Route::get("/cart/{id}",[UserProductController::class,'userProducts']);
+
+Route::get("/allproducts",[ProductController::class,'index']);
 Route::post("/cart",[UserProductController::class,'store']);
 Route::put("/cart/{id}",[UserProductController::class,'update']);
-Route::delete("/delete-from-cart/{userid}/{productid}",[UserProductController::class,'removeFromCart']);
+Route::middleware('auth:sanctum')->delete("/delete-from-cart/{userid}/{productid}",[UserProductController::class,'removeFromCart']);
 Route::put("/cart/update/{id}",[UserProductController::class,'update']);
 Route::resource('/categories',CategoryController::class);
 Route::resource('/products',ProductController::class);
 Route::delete("/clear-cart/{id}",[UserProductController::class,'destroy']);
+Route::delete("/products/{id}",[ProductController::class,'destroy']);
 Route::post('/order',[OrderController::class,'store']);
+Route::get('/orders/{id}',[OrderDetailsController::class,'showUserOrders']);
 Route::post('/order-details',[OrderDetailsController::class,'store']);
 Route::get('/productswithcart/{id}',[ProductController::class, 'allProducts']);
+Route::post("/admins/register",[AdminController::class,'register']);
+Route::post("/admins/login",[AdminController::class,'login']);
+Route::get("/admins/products",[AdminController::class,'getProducts']);
+
+
+
+Route::middleware('auth:sanctum')->group(function(){
+    Route::get("/cart/{id}",[UserProductController::class,'userProducts']);
+
+
+
+
+});
